requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-openai
chromadb
streamlit
streamlit-mic-recorder
beautifulsoup4
openai
python-dotenv
tiktoken
pyaudio
gTTS
SpeechRecognition
faiss-cpu
langchain-text-splitters
edge-tts
pydub
deepgram-sdk==3.2.0